refactor(): код перенесён в отдельный модуль
