+++ .github/workflows/python-package.yaml
@@ -5,9 +5,11 @@ name: Python package
 
 on:
   push:
-    branches: [ "master", "development" ]
+    branches:
+      - '**'
   pull_request:
-    branches: [ "master" ]
+    branches:
+      - 'master'
 
 jobs:
   build:
