+++ src/config.py
@@ -1,0 +1,6 @@
+MIN_PERIODS = 1
+MAX_PERIODS = 60
+MIN_AMOUNT = 10_000
+MAX_AMOUNT = 3_000_000
+MIN_RATE = 1
+MAX_RATE = 8
