+++ tests/test_route.py
@@ -1,5 +1,7 @@
 import unittest
-from app import app
+from http import HTTPStatus
+
+from src.app import app
 
 
 class TestDepositRoute(unittest.TestCase):
@@ -11,7 +13,7 @@ class TestDepositRoute(unittest.TestCase):
         response = self.app.get(
             f"/deposit?date={deposit_json['date']}&periods={deposit_json['periods']}&amount={deposit_json['amount']}&rate={deposit_json['rate']}"
         )
-        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.status_code, HTTPStatus.OK)
         self.assertIsInstance(response.json, dict)
 
     def test_invalid_deposit(self):
